Refactor: Lazy load PDF.js and remove client-only component
